main/requirements.txt
chinese_calendar==1.8.1
exchange_calendars==4.5
jqdatasdk==1.9.1
keyboard==0.13.5
openpyxl==3.1.2
pandas==2.1.1
pyzmq==25.1.1
